timedit_stress_project/requirements.txt
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
torch>=2.0
tqdm>=4.65
scikit-learn>=1.3
